Module Wu_Manber

module FirstMatch : sig ... end

An module for fuzzy searching the first match in a sequence

module Matcher : sig ... end

Collection of functors for creating matchers

module Patterns : sig ... end

Collection of modules types for patterns with different operations

module StringSearch : sig ... end

An module for fuzzy searching in strings

module Utils : sig ... end

Internal Utility Module

module WuManber : sig ... end

This module contains the main parts of the shift-or variant of the Wu and Manber algorithm