Module Wu_Manber.Patterns

Collection of modules types for patterns with different operations

module type Pattern = sig ... end

Basic pattern.

module type Elems = sig ... end

Some notion of characters.

Useful operators for patterns and elements

module type ElemsWithEquality = sig ... end

Some notion of characters and equality between characters.

module type ElemsIndexable = sig ... end

Characters can be enumerated from 0 to some n.

module type PatternWithElemEquality = sig ... end

Patterns with a notion of characters and a equality of characters operator.

module type PatternWithFoldRight = sig ... end

Patterns with a fold right operation.

module type PatternWithFoldLeft = sig ... end

Patterns with both fold right and fold left operations.

module type PatternWithIndexableElements = sig ... end

Patterns with a fold right operation where the elements can enumerated from 0 to some max_elem_index.