Module Wu_Manber.WuManber

This module contains the main parts of the shift-or variant of the Wu and Manber algorithm

module BitOps : sig ... end

A collection of bitwise operations used in the Wu and Manber algorithm.

module WuManber : sig ... end

Basic Wu and Manber algorithm.

module RightLeaningWuManber : sig ... end

Wu and Manber algorithm modified for right leaning matches.