Module WuManber.BitOps

A collection of bitwise operations used in the Wu and Manber algorithm.

val match_error : pattern_length:int -> Optint.Int63.t array -> int option

match_error ~pattern_length takes an array of bitvectors and returns Some n if theres a match with n errors, and None if there is no match. pattern_length must be less than or equal 63.

val is_match : pattern_length:int -> Optint.Int63.t array -> bool

is_match ~pattern_length takes an array of bitvectors and returns true if theres a match. pattern_length must be less than or equal 63.