Module WuManber.WuManber

Basic Wu and Manber algorithm.

val initial_bvs : k:int -> Optint.Int63.t array

initial_bv ~k creates a starting array of bitvectors used by the algorithm.

val next_bvs : mismatch:Optint.Int63.t -> Optint.Int63.t array -> Optint.Int63.t array

next_bvs ~mismatch bvs produces an updated bitvector array based on ~mismatch.